Generative_AI_for_Data/Fall_2024/Group_2_course_suggestion_chatbot_using_RAG/requirements.txt
langchain
pinecone-client
streamlit
openai
pypdf
python-dotenv
pandas